+++ src/views/noUpload/LocalNoUploadSummary.vue
@@ -2,20 +2,18 @@
   <div class="flex h-full w-full max-w-screen-xl flex-col items-center justify-center p-3 pr-5">
     <div class="flex h-full w-full flex-col">
       <span class="text-left text-lg font-medium"> Summary </span>
-      <!-- <span class="text-left">
-        We will use this to upload and edit your dataset on your Zenodo account.
-      </span> -->
 
-      <el-divider class="my-4"> </el-divider>
+      <el-divider class="my-2"> </el-divider>
 
       <div>
         <p class="mb-5">
-          A list of all the files and folders that we are going to upload to Zenodo is shown below.
-          Files that will be created by FAIRshare are highlighted in orange. You can click on any of
-          them to view their contents or open the file in your file browser.
+          You have selected to not upload your dataset on a repository. An overview of your dataset
+          is provided below. Metadata files that will be generated and included by FAIRshare (if
+          any) at your source folder selected in Step 1 are highlighted in orange. Click continue
+          below to generate these files.
         </p>
-        <div class="overflow-auto" :class="{ 'h-[200px]': finishedLoading }">
-          <fade-transition>
+        <div class="h-[200px] overflow-auto">
+          <transition name="fade" mode="out-in" appear>
             <el-tree-v2 v-if="finishedLoading" :data="fileData" :props="defaultProps">
               <template #default="{ node, data }">
                 <el-icon v-if="!node.isLeaf"><folder-icon /></el-icon>
@@ -62,7 +60,7 @@
                 :height="150"
               />
             </div>
-          </fade-transition>
+          </transition>
         </div>
         <el-drawer
           v-if="anyfilePreview"
@@ -108,35 +106,11 @@
       </div>
 
       <line-divider></line-divider>
-      <fade-transition>
-        <div v-if="ready">
-          <p v-if="validTokenAvailable" class="my-10 w-full text-center">
-            It looks like you have already connected your Zenodo account with FAIRshare. <br />
-            Click on the 'Start upload' button below to upload your dataset. <br />
-            <span class="text-xs">
-              Note that this will create a draft of the dataset on Zenodo that is only visible to
-              you. <br />
-              You will be prompted to Publish the dataset and make it openly available after the
-              upload.
-            </span>
-          </p>
-          <!-- show error message if token is not valid -->
-          <div v-else class="flex flex-col items-center justify-center py-10">
-            <p class="mb-5">
-              We couldn't find your Zenodo account details. Please click on the button below to
-              connect to your Zenodo account for uploading your dataset.
-            </p>
 
-            <ConnectZenodo :statusChangeFunction="showConnection"></ConnectZenodo>
-          </div>
-        </div>
-        <LoadingFoldingCube v-else></LoadingFoldingCube>
-      </fade-transition>
-
-      <fade-transition>
-        <div class="flex w-full flex-row justify-center space-x-4 py-2" v-if="finishedLoading">
+      <transition name="fade" mode="out-in" appear>
+        <div class="flex w-full flex-row justify-center space-x-4 py-2">
           <router-link
-            :to="`/datasets/${this.$route.params.datasetID}/${this.$route.params.workflowID}/zenodo/metadata`"
+            :to="`/datasets/${this.$route.params.datasetID}/${this.$route.params.workflowID}/selectDestination`"
             class=""
           >
             <button class="primary-plain-button">
@@ -144,35 +118,17 @@
             </button>
           </router-link>
 
-          <button
-            class="secondary-plain-button hidden"
-            @click="showFilePreview"
-            v-if="validTokenAvailable"
-          >
-            <el-icon><checked-icon /></el-icon>
-            View files ready for upload
-          </button>
-
-          <button
-            class="primary-button"
-            :disabled="disableContinue"
-            @click="uploadToZenodo"
-            v-if="validTokenAvailable"
-          >
-            Start upload
+          <button class="primary-button" @click="generateMetadataFiles">
+            Continue
             <el-icon> <d-arrow-right /> </el-icon>
           </button>
         </div>
-      </fade-transition>
+      </transition>
     </div>
-    <app-docs-link url="curate-and-share/zenodo-upload-summary" position="bottom-4" />
   </div>
 </template>
 
 <script>
-import LoadingFoldingCube from "@/components/spinners/LoadingFoldingCube";
-import ConnectZenodo from "@/components/serviceIntegration/ConnectZenodo";
-
 import { useDatasetsStore } from "@/store/datasets";
 import { useTokenStore } from "@/store/access.js";
 
@@ -185,8 +141,8 @@ import { marked } from "marked";
 import { v4 as uuidv4 } from "uuid";
 
 export default {
-  name: "ZenodoAccessToken",
-  components: { LoadingFoldingCube, ConnectZenodo },
+  name: "LocalNoUploadSummary",
+
   data() {
     return {
       datasetStore: useDatasetsStore(),
@@ -195,9 +151,6 @@ export default {
       folderPath: "",
       workflowID: this.$route.params.workflowID,
       workflow: {},
-      validTokenAvailable: false,
-      errorMessage: "",
-      zenodoAccessToken: "",
       ready: false,
       showFiles: "1",
       licenseData: "",
@@ -223,16 +176,6 @@ export default {
   },
   //el-tree-node__content
   computed: {
-    disableContinue() {
-      if (this.validTokenAvailable) {
-        return false;
-      }
-      if (!this.validTokenAvailable && this.zenodoAccessToken !== "") {
-        return false;
-      }
-      return true;
-    },
-
     anyfilePreview() {
       if (
         this.PreviewNewlyCreatedMetadataFile ||
@@ -385,8 +328,6 @@ export default {
       return response;
     },
     async showFilePreview() {
-      this.finishedLoading = false;
-
       this.fileData = [];
 
       this.tableDataRecord = [];
@@ -575,44 +516,10 @@ export default {
       }
     },
 
-    async checkToken(token) {
-      console.log(token);
-      const response = await this.tokens.getDepositions(token);
+    async generateMetadataFiles() {
+      const routerPath = `/datasets/${this.datasetID}/${this.workflowID}/localNoUpload/generate`;
 
-      if (response.status === 200) {
-        this.validTokenAvailable = true;
-        this.tokens.saveToken("zenodo", token);
-        return true;
-      } else if (response.status === 401) {
-        this.errorMessage =
-          "Invalid Zenodo access token. Please enter a valid Zenodo access token.";
-        this.validTokenAvailable = false;
-        return false;
-      } else {
-        this.errorMessage = "Something went wrong. Please try again.";
-        this.validTokenAvailable = false;
-        return false;
-      }
-    },
-    async uploadToZenodo() {
-      const routerPath = `/datasets/${this.datasetID}/${this.workflowID}/zenodo/upload`;
-      if (this.validTokenAvailable) {
-        this.$router.push({ path: routerPath });
-      } else {
-        const zenodoToken = this.zenodoAccessToken;
-        const res = await this.checkToken(zenodoToken);
-
-        if (res) {
-          this.$router.push({ path: routerPath });
-        }
-      }
-    },
-    async showConnection(status) {
-      console.log(status);
-      if (status === "connected") {
-        this.validTokenAvailable = true;
-      }
-      // this.uploadToZenodo(); console.log(this.workflow.licenseText)
+      this.$router.push({ path: routerPath });
     },
   },
   async mounted() {
@@ -624,14 +531,6 @@ export default {
     this.datasetStore.setCurrentStep(6);
 
     this.workflow.currentRoute = this.$route.path;
-
-    const validZenodoConnection = await this.tokens.verifyZenodoConnection();
-
-    if (validZenodoConnection) {
-      this.validTokenAvailable = true;
-    } else {
-      this.validTokenAvailable = false;
-    }
 
     this.showFilePreview();
   },