+++ src/views/project/CreateNewProjectConfirm.vue
@@ -7,7 +7,7 @@
 
       <line-divider> </line-divider>
 
-      <div class="flex flex-col items-center justify-center">
+      <div class="flex h-full flex-col items-center justify-center">
         <Vue3Lottie
           animationLink="https://assets9.lottiefiles.com/packages/lf20_bk87clix.json"
           :width="200"