+++ src/views/workflow/Code/ReviewStandards.vue
@@ -2,7 +2,8 @@
   <div class="flex h-full w-full max-w-screen-xl flex-col items-center justify-center p-3 pr-5">
     <div class="flex h-full w-full flex-col">
       <span class="text-left text-lg font-medium">
-        Verify if you have followed standard/best research software development practices
+        Verify if you have followed standard/best research software development practices required
+        to make your software FAIR
       </span>
 
       <el-divider class="my-4"> </el-divider>
@@ -11,7 +12,9 @@
         <!-- template start -->
         <div class="rounded-lg border-2 border-slate-100 p-4">
           <p class="pb-2">
-            Have you followed applicable high-level best practices for biomedical research software?
+            Have you followed applicable high-level standards and best practices for scientific
+            research software? Working from a version control system such as GitHub is especially
+            recommended.
           </p>
 
           <el-collapse v-model="activeNames" class="border-none">
@@ -38,7 +41,8 @@
         </div>
         <div class="rounded-lg border-2 border-slate-100 p-4">
           <p class="pb-2">
-            Have you followed applicable language-specific standards and best practices?
+            Have you followed applicable language-specific standards and best practices for your
+            code?
           </p>
 
           <el-collapse v-model="activeNames" class="border-none">
@@ -72,7 +76,8 @@
         </div>
         <div class="rounded-lg border-2 border-slate-100 p-4">
           <p class="pb-2">
-            Have you followed applicable domain-specific standards and best practices?
+            Have you followed applicable domain-specific standards and best practices for biomedical
+            research software?
           </p>
 
           <el-collapse v-model="activeNames" class="border-none">
@@ -86,21 +91,9 @@
                 research software:
                 <span
                   class="text-url"
-                  @click="openWebsite('https://doi.org/10.1016/j.cels.2018.03.014')"
-                >
-                  Practical Computational Reproducibility in the Life Sciences</span
-                >,
-                <span
-                  class="text-url"
                   @click="openWebsite('https://dx.doi.org/10.12688%2Ff1000research.10750.2')"
                 >
-                  General guidelines for biomedical software development </span
-                >,
-                <span
-                  class="text-url"
-                  @click="openWebsite('https://doi.org/10.3389/fgene.2014.00199')"
-                >
-                  On best practices in the development of bioinformatics software</span
+                  General guidelines for biomedical software development</span
                 >.
               </div>
             </el-collapse-item>
@@ -113,7 +106,10 @@
         </div>
         <div class="rounded-lg border-2 border-slate-100 p-4">
           <p class="pb-2">
-            Is your software documentation following applicable standards and best practices?
+            Is your software documentation following applicable standards and best practices? It is
+            especially recommemded that the following aspects are documented: inputs and outputs of
+            the software, parameters and data required to run the software, the standards applied,
+            and how to contribute to the software.
           </p>
 
           <el-collapse v-model="activeNames" class="border-none">
@@ -123,7 +119,8 @@
               </template>
 
               <div class="px-3 py-2">
-                We refer to the following literature:
+                We refer to the following literature for additional information on documenting your
+                software:
                 <span
                   class="text-url"
                   @click="openWebsite('https://dx.doi.org/10.1371%2Fjournal.pcbi.1006561')"